# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-MKITIPONG",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/emails/{id}": {
            "get": {
                "tags": [
                    "email"
                ],
                "summary": "email information",
                "description": "Returns email data",
                "operationId": "App\\Http\\Controllers\\EmailController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "email id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "email"
                ],
                "summary": "update email",
                "description": "",
                "operationId": "App\\Http\\Controllers\\EmailController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "email id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "people_id",
                                    "name",
                                    "address",
                                    "other_details"
                                ],
                                "properties": {
                                    "people_id": {
                                        "type": "integer",
                                        "format": "people_id",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "user1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "address",
                                        "example": "user1@mail.com"
                                    },
                                    "other_details": {
                                        "type": "string",
                                        "format": "other_details",
                                        "example": "detail"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "email"
                ],
                "summary": "delete email",
                "description": "delete data",
                "operationId": "App\\Http\\Controllers\\EmailController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "email id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/eucations/{id}": {
            "get": {
                "tags": [
                    "eucation"
                ],
                "summary": "eucation information",
                "description": "Returns eucation data",
                "operationId": "App\\Http\\Controllers\\EucationController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "eucation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "eucation"
                ],
                "summary": "update eucation",
                "description": "",
                "operationId": "App\\Http\\Controllers\\EucationController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "eucation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "people_id",
                                    "level",
                                    "degree",
                                    "branch",
                                    "faculty",
                                    "academy",
                                    "gpa",
                                    "graduation_date",
                                    "other_details"
                                ],
                                "properties": {
                                    "people_id": {
                                        "type": "integer",
                                        "format": "people_id",
                                        "example": "1"
                                    },
                                    "level": {
                                        "type": "string",
                                        "format": "level",
                                        "example": "level"
                                    },
                                    "degree": {
                                        "type": "string",
                                        "format": "degree",
                                        "example": "degree"
                                    },
                                    "branch": {
                                        "type": "string",
                                        "format": "branch",
                                        "example": "branch"
                                    },
                                    "faculty": {
                                        "type": "string",
                                        "format": "faculty",
                                        "example": "faculty"
                                    },
                                    "academy": {
                                        "type": "string",
                                        "format": "academy",
                                        "example": "academy"
                                    },
                                    "gpa": {
                                        "type": "string",
                                        "format": "gpa",
                                        "example": "gpa"
                                    },
                                    "graduation_date": {
                                        "type": "date",
                                        "format": "graduation_date",
                                        "example": "2021-01-01"
                                    },
                                    "other_details": {
                                        "type": "string",
                                        "format": "other_details",
                                        "example": "detail"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "eucation"
                ],
                "summary": "delete eucation",
                "description": "delete data",
                "operationId": "App\\Http\\Controllers\\EucationController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "eucation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/specials": {
            "get": {
                "tags": [
                    "special"
                ],
                "summary": "special information",
                "description": "Returns special data",
                "operationId": "App\\Http\\Controllers\\SpecialController::index",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "special"
                ],
                "summary": "add special",
                "description": "",
                "operationId": "App\\Http\\Controllers\\SpecialController::store",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "other_details"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "user1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "address",
                                        "example": "user1@mail.com"
                                    },
                                    "other_details": {
                                        "type": "string",
                                        "format": "other_details",
                                        "example": "detail"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/specials/del/{id}": {
            "put": {
                "tags": [
                    "special"
                ],
                "summary": "disable status special",
                "description": "",
                "operationId": "App\\Http\\Controllers\\SpecialController::del",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "special id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}